# API_Documentation.yaml
openapi: '3.0.2'
info:
  title: ISS Capstone API
  version: '1.0'
  description: 
    This is the API documentation for the ISS Capstone project.
    For more information, contact the developers.
servers:
  - url: https://api.server.test/v1
paths:
...
  /Devices/:
    get:
      summary: Get all registered devices
      tags:
        - Devices
      responses:
        '200':
          description: Returns all registered devices
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  example: 
                    _id: 1
                    Name: "Front-door device"
                    Description: "This is device at the front door"
                    Status: "ACTIVE"
                    AuthToken: "1234567890"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Unauthorized"

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Internal server error"
      security:
        - bearerAuth: []
  /Devices/{_id}:
    get:
      summary: Get device by ID
      tags:
        - Devices
      parameters:
        - in: path
          name: _id
          schema:
            type: integer
          required: true
          description: Numeric ID of the device to get
      responses:
        '200':
          description: Get all registered devices
          content:
            application/json:
              schema:
                type: object
                example:
                    _id: 1
                    Name: "Front-door device"
                    Description: "This is device at the front door"
                    Status: "ACTIVE"
                    AuthToken: "1234567890"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Unauthorized"
        "404":
          description: "Device not found"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Device not found"

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Internal server error"
      security:
        - bearerAuth: []
  /Devices/register:
    post:
      tags:
      - Devices
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Name:
                  type: string
                  example: "Front-door device"
                Description:
                  type: string
                  example: "This is device at the front door"
                Status:
                  type: string
                  example: "ACTIVE"
                AuthToken:
                  type: string
                  example: "1234567890"
              required:
                - Name
                - Description
                - Status
                - AuthToken
      responses:
        '201':
          description: Device added successfully
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: false
                  message: "Device added successfully!"
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Unauthorized"
        '400':
          description: "Bad request"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Detail of missing or mailformed request body"
        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Internal server error"
      security:
          - bearerAuth: []
...
  /Users:
    get:
      summary: Get all  users
      tags:
        - Users
      responses:
        '200':
          description: Returns all registered users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  example: 
                    _id: 1
                    PetId: 1
                    EmployeeId: "123456789"
                    Score: 9999
                    Credit: 9999
  
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Unauthorized"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Internal server error"
      security:
        - bearerAuth: []
...
  /Badges:
    get:
      summary: Get all badges
      tags:
        - Badges
      responses:
        '200':
          description: Returns all badges in the data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  example: 
                    _id: 1
                    Title: "Badge 1"
                    Description: "This is badge 1"
                    ImageURL: "localhost:8080/image1"
                    
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Unauthorized"

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Internal server error"
      security:
        - bearerAuth: []
  /Badges/{_id}:
    get:
      summary: Get badge by ID
      tags:
        - Badges
      parameters:
        - in: path
          name: _id
          schema:
            type: integer
          required: true
          description: Numeric ID of the badge to get
      responses:
        '200':
          description: Get all registered badges
          content:
            application/json:
              schema:
                type: object
                example:
                    _id: 1
                    Title: "Badge 1"
                    Description: "This is badge 1"
                    ImageURL: "localhost:8080/image1"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Unauthorized"
        "404":
          description: "Badge not found"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Badge not found"

        '500':
          description: "Internal server error"
          content:
            application/json:
              schema:
                type: object
                example: 
                  error: true
                  message: "Internal server error"
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT